+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/), from version **1.3.0** on.  
   
-## [1.5.1]
+## [1.6.0]
 
 ### Changed
 - WtsSession:
@@ -12,16 +12,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - `Remove-Service` Cmdlet.
+- `Get-ServiceSecurity` Cmdlet.
+- `New-ServiceAccessRule` Cmdlet.
+- `New-ServiceAuditRule` Cmdlet.
 - `NativeException` exception on both core, and main libraries to wrap Win32 errors.
 - Added help message on all parameters.
 - Added the `ValidateFileExists` attribute to validate if single files exists.
+- Added `Format.ps1xml` for the new type `WindowsUtils.AccessControl.ServiceSecurity`.
+- Added type support on `Types.ps1xml` for `WindowsUtils.AccessControl.ServiceSecurity`.
 
 ### Bugs
 - Get-ResourceMessageTable:
   - "s" was being used as a format specifier, which was breaking the message strings. Replaced with "S".  
   - Trimmed output string on `Message` to remove the extra line feed.
 
-## [1.5.0] - 2023-01-18
+## [1.5.1] - 2023-01-18
 
 ### Changed
 - Structure: